+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to
 
  - Account payload size of the deferred forward TSN messages.
 
+### Added
+
+ - Added Verification Tag validation on incoming packets as per RFC 9260 Section 8.5.
+
 ## 0.1.13 - 2026-05-20
 
 ### Fixed